Rename NewMainActivity to MainActivity, document and clean up MainActivity
